requirements.txt
ipykernel
pandas
matplotlib
scipy
statsmodels
sympy
seaborn
# tensorflow
tensorflow[and-cuda]
pydot
graphviz
jinja2
scikit-learn
fastapi
uvicorn
pillow
python-multipart